pyproject.toml
[project]
name = "watchdog-analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.10,<3.15"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "aiokafka (>=0.13.0,<0.14.0)",
    "aiogram (>=3.27.0,<4.0.0)",
    "aiohttp (>=3.13.5,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"